move views to own dir
