+++ client/src/views/Replay.tsx
@@ -1,9 +1,9 @@
 import { Alert, Button, Card, FileButton, Group, LoadingOverlay, Stack, Title, Tooltip } from "@mantine/core";
 import { IconFileUpload, IconInfoCircle } from "@tabler/icons-react";
 import { useRef, useState } from "react";
-import { Game } from "./components/Game";
-import { useReplay } from "./hooks/useReplay";
 import { FormattedMessage, useIntl } from "react-intl";
+import { Game } from "../components/Game";
+import { useReplay } from "../hooks/useReplay";
 
 enum ReplayState {
     ProvideFile,
